+++ DelvUI/Interface/Nameplates/NameplatesHud.cs
@@ -182,7 +182,8 @@ namespace DelvUI.Interface.Nameplates
             Vector3 cameraPos = camera.Object.Position;
 
             BGCollisionModule* collisionModule = Framework.Instance()->BGCollisionModule;
-            int* flags = stackalloc int[] { 0x2000, 0, 0x2000, 0 };
+            int flag = Config.RaycastFlag();
+            int* flags = stackalloc int[] { flag, 0, flag, 0 };
             bool obstructed = false;
 
             // simple mode